move trajectory props to global props file
