+++ molecularnodes/addon.py
@@ -65,7 +65,7 @@ def register():
     bpy.types.Object.mn = PointerProperty(type=props.MolecularNodesObjectProperties)  # type: ignore
     bpy.types.Scene.mn = PointerProperty(type=props.MolecularNodesSceneProperties)  # type: ignore
     bpy.types.Object.mn_trajectory_selections = CollectionProperty(  # type: ignore
-        type=entities.trajectory.props.TrajectorySelectionItem  # type: ignore
+        type=props.TrajectorySelectionItem  # type: ignore
     )
